// src/interfaces/Enums.ts
export enum Path {
  Abundance = "The Abundance",
  Destruction = "The Destruction",
  Erudition = "The Erudition",
  Harmony = "The Harmony",
  Hunt = "The Hunt",
  Nihility = "The Nihility",
  Preservation = "The Preservation"
}

export enum Element {
  Fire = "Fire",
  Ice = "Ice",
  Imaginary = "Imaginary",
  Lightning = "Lightning",
  Physical = "Physical",
  Quantum = "Quantum",
  Wind = "Wind"
}

export enum Sex {
  Male = "Male",
  Female = "Female"
}

export enum Faction {
  AstralExpress = "Astral Express",
  Belobog = "Belobog",
  HertaSpaceStation = "Herta Space Station",
  InterastralPeaceCorporation = "Interastral Peace Corporation",
  IntergalacticMerchantGuild = "Intergalactic Merchant Guild",
  SilvermaneGuards = "Silvermane Guards",
  StellaronHunters = "Stellaron Hunters",
  Wildfire = "Wildfire",
  XianzhouLuofu = "Xianzhou Luofu"
}